requirements.txt
asgiref==3.8.1
camelot-py==0.11.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
ghostscript==0.7
Levenshtein==0.26.1
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pdfminer.six==20240706
pycparser==2.22
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2024.2
RapidFuzz==3.10.1
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
tabula==1.0.5
tabula-py==2.10.0
tabulate==0.9.0
tzdata==2024.2
